refactor(wizard): split flow.py (2813 lines) into three focused modules
Extract UI primitives into _ui.py, integration configurators into
_integration_configurators.py, and reduce flow.py to orchestration only
(run_wizard, _run_cli_llm_onboarding, _credential_line_for_saved_summary).
Remove 27 lazy-import stub functions replaced by direct imports.
Update all test monkeypatches to target the correct defining module.
