+++ tests/cli/wizard/test_model_selection.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from app.cli.wizard import flow
+from app.cli.wizard import _ui, flow
 from app.cli.wizard.config import PROVIDER_BY_VALUE
 
 
@@ -29,7 +29,7 @@ def _wire_prompts(
         m.ask.return_value = next(text_iter)
         return m
 
-    monkeypatch.setattr(flow, "select_prompt", _mock_select)
+    monkeypatch.setattr(_ui, "select_prompt", _mock_select)
     monkeypatch.setattr(flow.questionary, "text", _mock_text)
 
 
@@ -37,7 +37,7 @@ def test_choose_model_returns_curated_default(monkeypatch: pytest.MonkeyPatch) -
     provider = PROVIDER_BY_VALUE["anthropic"]
     _wire_prompts(monkeypatch, select_values=[provider.default_model])
 
-    model = flow._choose_model(provider, default=provider.default_model)
+    model = _ui._choose_model(provider, default=provider.default_model)
 
     assert model == provider.default_model
 
@@ -53,13 +53,13 @@ def test_choose_model_offers_full_curated_list(monkeypatch: pytest.MonkeyPatch) 
         m.ask.return_value = provider.default_model
         return m
 
-    monkeypatch.setattr(flow, "select_prompt", _mock_select)
+    monkeypatch.setattr(_ui, "select_prompt", _mock_select)
 
-    flow._choose_model(provider, default="")
+    _ui._choose_model(provider, default="")
 
     expected_curated = [opt.value for opt in provider.models]
     assert captured["values"][:-1] == expected_curated
-    assert captured["values"][-1] == flow._CUSTOM_MODEL_SENTINEL
+    assert captured["values"][-1] == _ui._CUSTOM_MODEL_SENTINEL
 
 
 def test_choose_model_preserves_saved_model_not_in_curated(
@@ -75,9 +75,9 @@ def test_choose_model_preserves_saved_model_not_in_curated(
         m.ask.return_value = "my-tuned-gpt"
         return m
 
-    monkeypatch.setattr(flow, "select_prompt", _mock_select)
+    monkeypatch.setattr(_ui, "select_prompt", _mock_select)
 
-    model = flow._choose_model(provider, default="my-tuned-gpt")
+    model = _ui._choose_model(provider, default="my-tuned-gpt")
 
     assert model == "my-tuned-gpt"
     assert "my-tuned-gpt" in captured["values"]
@@ -87,11 +87,11 @@ def test_choose_model_accepts_custom_entry(monkeypatch: pytest.MonkeyPatch) -> N
     provider = PROVIDER_BY_VALUE["anthropic"]
     _wire_prompts(
         monkeypatch,
-        select_values=[flow._CUSTOM_MODEL_SENTINEL],
+        select_values=[_ui._CUSTOM_MODEL_SENTINEL],
         text_values=["claude-future-preview"],
     )
 
-    model = flow._choose_model(provider, default=provider.default_model)
+    model = _ui._choose_model(provider, default=provider.default_model)
 
     assert model == "claude-future-preview"
 
@@ -101,6 +101,6 @@ def test_choose_model_works_for_cli_provider(monkeypatch: pytest.MonkeyPatch) ->
     provider = PROVIDER_BY_VALUE["codex"]
     _wire_prompts(monkeypatch, select_values=["gpt-5.4"])
 
-    model = flow._choose_model(provider, default="")
+    model = _ui._choose_model(provider, default="")
 
     assert model == "gpt-5.4"
